+++ control/tests/xferfcn_input_test.py
@@ -54,15 +54,15 @@ cases = {
 
 
 @pytest.mark.parametrize("dtype",
-                         [np.int, np.int8, np.int16, np.int32, np.int64,
-                          np.float, np.float16, np.float32, np.float64,
+                         [int, np.int8, np.int16, np.int32, np.int64,
+                          float, np.float16, np.float32, np.float64,
                           np.longdouble])
 @pytest.mark.parametrize("num, fun", cases.values(), ids=cases.keys())
 def test_clean_part(num, fun, dtype):
     """Test clean part for various inputs"""
     numa = fun(dtype, num)
     num_ = _clean_part(numa)
-    ref_ = np.array(num, dtype=np.float, ndmin=3)
+    ref_ = np.array(num, dtype=float, ndmin=3)
 
     assert isinstance(num_, list)
     assert np.all([isinstance(part, list) for part in num_])
